requirements.txt
flask-executor==0.9.3
aniso8601==8.0.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
docopt==0.6.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
hdfs==2.5.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
neo4j==4.0.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.4.2
pykerberos==1.2.1
pytz==2020.1
requests==2.24.0
requests-kerberos==0.12.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
gunicorn==20.0.4
python-keycloak==0.21.0
flask-restx==0.2.0